Tidy

+++ data/model/test/MockWaveModel.cpp
@@ -67,17 +67,17 @@ MockWaveModel::generate(Sort sort, int length) const
 
     for (int i = 0; i < length; ++i) {
 
-	float v = 0.f;
+	double v = 0.0;
 	
 	switch (sort) {
-	case DC: v = 1.f; break;
-	case Sine: v = (float)sin((2.0 * M_PI / 8.0) * i); break;
-	case Cosine: v = (float)cos((2.0 * M_PI / 8.0) * i); break;
+	case DC: v = 1.0; break;
+	case Sine: v = sin((2.0 * M_PI / 8.0) * i); break;
+	case Cosine: v = cos((2.0 * M_PI / 8.0) * i); break;
 	case Nyquist: v = (i % 2) * 2 - 1; break;
-	case Dirac: v = (i == 0) ? 1.f : 0.f; break;
+	case Dirac: v = (i == 0) ? 1.0 : 0.0; break;
 	}
 
-	data.push_back(v);
+	data.push_back(float(v));
     }
 
     return data;
